+++ app/src/main/java/com/tajmoti/tulip/ui/player/streams/AndroidStreamsViewModel.kt
@@ -1,18 +1,17 @@
 package com.tajmoti.tulip.ui.player.streams
 
-import androidx.lifecycle.*
+import androidx.lifecycle.viewModelScope
 import com.tajmoti.libtulip.repository.StreamsRepository
-import com.tajmoti.libtulip.service.*
+import com.tajmoti.libtulip.service.LanguageMappingStreamService
+import com.tajmoti.libtulip.service.VideoDownloadService
 import com.tajmoti.libtulip.ui.streams.StreamsViewModel
 import com.tajmoti.libtulip.ui.streams.StreamsViewModelImpl
 import com.tajmoti.tulip.ui.DelegatingViewModel
-import com.tajmoti.tulip.ui.player.VideoPlayerActivityArgs
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class AndroidStreamsViewModel @Inject constructor(
-    savedStateHandle: SavedStateHandle,
     downloadService: VideoDownloadService,
     extractionService: StreamsRepository,
     streamService: LanguageMappingStreamService
@@ -21,7 +20,6 @@ class AndroidStreamsViewModel @Inject constructor(
         downloadService,
         extractionService,
         streamService,
-        VideoPlayerActivityArgs.fromSavedStateHandle(savedStateHandle).streamableKey,
         viewModelScope
     )
 }