+++ libtulip/src/main/java/com/tajmoti/libtulip/ui/player/VideoPlayerViewModelImpl.kt
@@ -1,10 +1,15 @@
 package com.tajmoti.libtulip.ui.player
 
 import com.tajmoti.commonutils.map
+import com.tajmoti.libtulip.misc.NetworkResult
+import com.tajmoti.libtulip.model.info.TulipSeasonInfo
+import com.tajmoti.libtulip.model.key.EpisodeKey
 import com.tajmoti.libtulip.model.key.StreamableKey
 import com.tajmoti.libtulip.model.subtitle.SubtitleInfo
+import com.tajmoti.libtulip.repository.HostedTvDataRepository
 import com.tajmoti.libtulip.repository.PlayingHistoryRepository
 import com.tajmoti.libtulip.repository.SubtitleRepository
+import com.tajmoti.libtulip.repository.TmdbTvDataRepository
 import com.tajmoti.libtulip.service.*
 import com.tajmoti.libtulip.ui.doCancelableJob
 import com.tajmoti.libtvprovider.*
@@ -12,19 +17,51 @@ import kotlinx.coroutines.*
 import kotlinx.coroutines.flow.*
 import java.io.File
 
+@OptIn(ExperimentalCoroutinesApi::class)
 class VideoPlayerViewModelImpl constructor(
     private val subtitleRepository: SubtitleRepository,
     private val playingHistoryRepository: PlayingHistoryRepository,
+    private val tmdbTvDataRepository: TmdbTvDataRepository,
+    private val hostedTvDataRepository: HostedTvDataRepository,
     private val subDirectory: File,
     private val viewModelScope: CoroutineScope,
-    private val streamableKey: StreamableKey
+    streamableKeyInitial: StreamableKey
 ) : VideoPlayerViewModel {
+    override val streamableKey = MutableStateFlow(streamableKeyInitial)
+
+    override val episodeList = streamableKey
+        .flatMapLatest { getSeasonFlow(it) }
+        .map { it?.data?.episodes }
+        .stateIn(viewModelScope, SharingStarted.Eagerly, null)
+
+    private fun getSeasonFlow(
+        streamableKey: StreamableKey
+    ): Flow<NetworkResult<out TulipSeasonInfo>?> {
+        return when (streamableKey) {
+            is EpisodeKey.Tmdb -> tmdbTvDataRepository.getSeasonAsFlow(streamableKey.seasonKey)
+            is EpisodeKey.Hosted -> hostedTvDataRepository.getSeasonAsFlow(streamableKey.seasonKey)
+            else -> flowOf(null)
+        }
+    }
+
+    override val nextEpisode = combine(episodeList, streamableKey) { a, b -> a to b }
+        .map { (episodes, streamableKey) ->
+            val currentEpisodeIndex = episodes
+                ?.indexOfFirst { episode -> episode.key == streamableKey }
+                ?: return@map null
+            if (currentEpisodeIndex == episodes.size - 1)
+                return@map null
+            episodes[currentEpisodeIndex + 1].key
+        }
+        .stateIn(viewModelScope, SharingStarted.Eagerly, null)
 
     /**
      * State of subtitle list loading
      */
-    private val loadingSubtitlesState = loadSubtitlesList(streamableKey)
+    private val loadingSubtitlesState = streamableKey
+        .flatMapLatest { loadSubtitlesList(it) }
         .stateIn(viewModelScope, SharingStarted.Eagerly, SubtitleListLoadingState.Loading)
+
 
     /**
      * State of downloading of selected subtitles
@@ -35,18 +72,14 @@ class VideoPlayerViewModelImpl constructor(
 
     override val loadingSubtitles = MutableStateFlow(false)
     override val downloadingSubtitleFile = MutableStateFlow(false)
-    override val subtitleList = loadingSubtitlesState.map(viewModelScope) {
-        (it as? SubtitleListLoadingState.Success)?.subtitles ?: emptyList()
-    }
-    override val subtitlesReadyToSelect = loadingSubtitlesState.map(viewModelScope) {
-        it is SubtitleListLoadingState.Success
-    }
-    override val downloadingError = subtitleDownloadState.map(viewModelScope) {
-        it is SubtitleDownloadingState.Error
-    }
-    override val subtitleFile = subtitleDownloadState.map(viewModelScope) {
-        (it as? SubtitleDownloadingState.Success)?.subtitles
-    }
+    override val subtitleList = loadingSubtitlesState
+        .map(viewModelScope) { (it as? SubtitleListLoadingState.Success)?.subtitles ?: emptyList() }
+    override val subtitlesReadyToSelect = loadingSubtitlesState
+        .map(viewModelScope) { it is SubtitleListLoadingState.Success }
+    override val downloadingError = subtitleDownloadState
+        .map(viewModelScope) { it is SubtitleDownloadingState.Error }
+    override val subtitleFile = subtitleDownloadState
+        .map(viewModelScope) { (it as? SubtitleDownloadingState.Success)?.subtitles }
 
     /**
      * Currently attached media player.
@@ -56,26 +89,33 @@ class VideoPlayerViewModelImpl constructor(
     /**
      * State of the currently attached media player.
      */
-    @OptIn(ExperimentalCoroutinesApi::class)
-    private val mediaPlayerState = attachedMediaPlayer.flatMapLatest {
-        it?.state ?: flowOf(MediaPlayerHelper.State.Idle)
-    }.stateIn(viewModelScope, SharingStarted.Eagerly, MediaPlayerHelper.State.Idle)
+    private val mediaPlayerState = attachedMediaPlayer
+        .flatMapLatest {
+            it?.state ?: flowOf(MediaPlayerHelper.State.Idle)
+        }
+        .stateIn(viewModelScope, SharingStarted.Eagerly, MediaPlayerHelper.State.Idle)
+
     override val isPlaying = mediaPlayerState.map(viewModelScope) {
         it is MediaPlayerHelper.State.Playing
     }
-    override val showPlayButton = mediaPlayerState.map(viewModelScope) {
-        when (it) {
-            is MediaPlayerHelper.State.Buffering -> VideoPlayerViewModel.PlayButtonState.HIDE
-            is MediaPlayerHelper.State.Error -> VideoPlayerViewModel.PlayButtonState.HIDE
-            is MediaPlayerHelper.State.Idle -> VideoPlayerViewModel.PlayButtonState.HIDE
-            is MediaPlayerHelper.State.Paused -> VideoPlayerViewModel.PlayButtonState.SHOW_PLAY
-            is MediaPlayerHelper.State.Playing -> VideoPlayerViewModel.PlayButtonState.SHOW_PAUSE
-            is MediaPlayerHelper.State.Finished -> VideoPlayerViewModel.PlayButtonState.SHOW_PLAY
+
+    override val showPlayButton = mediaPlayerState
+        .map(viewModelScope) {
+            when (it) {
+                is MediaPlayerHelper.State.Buffering -> VideoPlayerViewModel.PlayButtonState.HIDE
+                is MediaPlayerHelper.State.Error -> VideoPlayerViewModel.PlayButtonState.HIDE
+                is MediaPlayerHelper.State.Idle -> VideoPlayerViewModel.PlayButtonState.HIDE
+                is MediaPlayerHelper.State.Paused -> VideoPlayerViewModel.PlayButtonState.SHOW_PLAY
+                is MediaPlayerHelper.State.Playing -> VideoPlayerViewModel.PlayButtonState.SHOW_PAUSE
+                is MediaPlayerHelper.State.Finished -> VideoPlayerViewModel.PlayButtonState.SHOW_PLAY
+            }
         }
-    }
-    override val buffering = mediaPlayerState.map(viewModelScope) { state ->
-        (state as? MediaPlayerHelper.State.Buffering)?.percent
-    }
+
+    override val buffering = mediaPlayerState
+        .map(viewModelScope) { state ->
+            (state as? MediaPlayerHelper.State.Buffering)?.percent
+        }
+
     override val position = mediaPlayerState
         .map { state ->
             when (state) {
@@ -91,10 +131,10 @@ class VideoPlayerViewModelImpl constructor(
         .stateIn(viewModelScope, SharingStarted.Eagerly, null)
 
     /**
-     * Playing progress (as a real number from 0.0 to 1.0)
+     * Playing progress (as a real number from 0.0 to 1.0) or null if nothing is currently playing.
      */
     val progress = mediaPlayerState
-        .mapNotNull { state ->
+        .map { state ->
             when (state) {
                 is MediaPlayerHelper.State.Buffering -> state.position
                 is MediaPlayerHelper.State.Error -> null
@@ -104,8 +144,8 @@ class VideoPlayerViewModelImpl constructor(
                 is MediaPlayerHelper.State.Finished -> null
             }
         }
-        .filter { isValidPosition(it) }
-        .map { it.fraction }
+        .filter { it?.let { isValidPosition(it) } ?: true }
+        .map { it?.fraction }
 
     private fun isValidPosition(it: Position): Boolean {
         return (it.timeMs > 0 && it.fraction > 0.0f)
@@ -115,19 +155,18 @@ class VideoPlayerViewModelImpl constructor(
      * Persisted playing position, should be used only to resume playback state
      * the first time the media is loaded.
      */
-    private val persistedPlayingProgress = playingHistoryRepository
-        .getLastPlayedPosition(streamableKey)
-        .map { it?.progress }
+    private val persistedPlayingProgress = streamableKey
+        .flatMapLatest {
+            playingHistoryRepository
+                .getLastPlayedPosition(it)
+                .map { position -> it to position?.progress }
+        }
 
-    override val lastValidPosition = position.mapNotNull { it?.timeMs }
-        .stateIn(viewModelScope, SharingStarted.Eagerly, 0L)
+    override val isDonePlaying = mediaPlayerState
+        .map(viewModelScope) { state -> state is MediaPlayerHelper.State.Finished }
 
-    override val isDonePlaying = mediaPlayerState.map(viewModelScope) { state ->
-        state is MediaPlayerHelper.State.Finished
-    }
-    override val isError = mediaPlayerState.map(viewModelScope) { state ->
-        state is MediaPlayerHelper.State.Error
-    }
+    override val isError = mediaPlayerState
+        .map(viewModelScope) { state -> state is MediaPlayerHelper.State.Error }
 
     /**
      * State of the interactive subtitle synchronization mechanism.
@@ -136,9 +175,8 @@ class VideoPlayerViewModelImpl constructor(
      */
     private val subSyncState = MutableStateFlow<SubtitleSyncState?>(null)
 
-    override val subtitleOffset = subSyncState.map(viewModelScope) {
-        (it as? SubtitleSyncState.OffsetUsed)?.offsetMs ?: 0L
-    }
+    override val subtitleOffset = subSyncState
+        .map(viewModelScope) { (it as? SubtitleSyncState.OffsetUsed)?.offsetMs ?: 0L }
 
     /**
      * Represents fetching of the selected subtitles.
@@ -156,9 +194,11 @@ class VideoPlayerViewModelImpl constructor(
     @OptIn(FlowPreview::class)
     private fun persistPlayingPosition() {
         viewModelScope.launch {
-            progress.sample(PLAY_POSITION_SAMPLE_PERIOD_MS).collect {
-                playingHistoryRepository.setLastPlayedPosition(streamableKey, it)
-            }
+            progress.sample(PLAY_POSITION_SAMPLE_PERIOD_MS)
+                .filterNotNull()
+                .collect {
+                    playingHistoryRepository.setLastPlayedPosition(streamableKey.value, it)
+                }
         }
     }
 
@@ -167,6 +207,10 @@ class VideoPlayerViewModelImpl constructor(
         val subtitles = subtitleRepository.fetchAvailableSubtitles(id)
             .getOrElse { emit(SubtitleListLoadingState.Error); return@flow }
         emit(SubtitleListLoadingState.Success(subtitles))
+    }
+
+    override fun goToNextEpisode() {
+        nextEpisode.value?.let { streamableKey.value = it }
     }
 
     override fun onMediaAttached(media: MediaPlayerHelper) {
@@ -179,14 +223,12 @@ class VideoPlayerViewModelImpl constructor(
      * and restores it to the media player if it is still attached.
      */
     private suspend fun restorePlayerPositionOrProgress() {
-        val lastPosition = lastValidPosition.value.takeIf { it != 0L }
-        val persistedProgress = persistedPlayingProgress.first()
-        val player = attachedMediaPlayer.value ?: return
-        if (lastPosition != null) {
-            player.time = lastPosition
-        } else if (persistedProgress != null) {
+        val (key, persistedProgress) = persistedPlayingProgress.first()
+        persistedProgress ?: return
+        val player = attachedMediaPlayer.value
+            ?: return
+        if (key == streamableKey.value)
             player.progress = persistedProgress
-        }
     }
 
     override fun onMediaDetached() {