+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - *Resume playing* feature for TV shows
+- *Next episode* button in video player for TV shows
 ### Changed
 - Increase thumb size on seekbar in video player
 - Search results now appear as they load instead of waiting for the last one before they are shown